refactor: improve core-utils folder structure
Improves the folder structure within the core-utils to make it more
intuitive. core-utils is now split into three primary folders, "common",
"external", and "optimism". "common" includes TypeScript/JavaScript
utilities not necessarily meant to be used with any particular package
(like hex string utilities). "external" includes utilities that are
meant to be used in conjunction with some other particular piece of
tooling (ethers provider utilities, geth typings). "optimism" includes
any utilities that are specifically meant to be used in the context of
Optimism (like address aliasing utils, batch encoding utils).
