+++ .changeset/poor-spies-provide.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/core-utils': patch
+---
+
+Refactor folder structure of @eth-optimism/core-utils.
