+++ packages/core-utils/src/common/index.ts
@@ -1,3 +1,9 @@
+/**
+ * Common JavaScript/TypeScript utilities
+ */
+
+export * from './basic-types'
+export * from './bn'
 export * from './hex-strings'
 export * from './misc'
 export * from './test-utils'
