+++ packages/core-utils/src/external/bcfg/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Utilities related to BCFG
+ */
+
+export * from './bcfg-types'
