+++ packages/core-utils/src/external/ethers/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Utilities that extend or enhance the ethers.js library
+ */
+
+export * from './fallback-provider'
