+++ packages/core-utils/src/external/geth/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Utilities related to go-ethereum (Geth)
+ */
+
+export * from './geth-types'
