+++ packages/core-utils/src/index.ts
@@ -1,10 +1,3 @@
-export * from './coders'
 export * from './common'
-export * from './watcher'
-export * from './l2context'
-export * from './types'
-export * from './fees'
-export * from './provider'
-export * from './alias'
-export * from './types'
-export * from './bn'
+export * from './external'
+export * from './optimism'
