+++ packages/core-utils/src/optimism/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Utils specifically related to Optimism.
+ */
+
+export * from './alias'
+export * from './batch-encoding'
+export * from './fees'
+export * from './l2context'
+export * from './rollup-types'
+export * from './watcher'
