--- packages/core-utils/src/types/index.ts
@@ -1,4 +1,0 @@
-export * from './geth'
-export * from './bcfg'
-export * from './rollup'
-export * from './basic'
