+++ packages/core-utils/test/alias.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from './setup'
-import { applyL1ToL2Alias, undoL1ToL2Alias } from '../src/alias'
+import { applyL1ToL2Alias, undoL1ToL2Alias } from '../src'
 
 describe('address aliasing utils', () => {
   describe('applyL1ToL2Alias', () => {
